Add `FromList`
Problem: we lack a generic function for constructing an object from
list, while `toList` is present.

Solution: add `fromList`.

The sad point here is that we cannot define it opposite to `toList`
because for e.g. `Map` our `toList` returns only values, unfortunatelly.
(Maybe this is worth changing one day?)

So `fromList` is opposite to something intermediate between `toList` and
`toPairs`, and so I put it to a separate typeclass.